got an alternate ui working for the image uploads
